Updates folder structure and adds basic positional styling to event calendar
