+++ src/components/Day/Day.module.scss
@@ -1,0 +1,7 @@
+.day {
+  width: 152px;
+  height: 200px;
+  border: 2px solid black;
+  padding: 8px 12px;
+  text-align: right;
+}
