+++ src/components/Day/Day.tsx
@@ -1,9 +1,11 @@
+import styles from "./Day.module.scss";
+
 interface DayProps {
   date: string;
 }
 
 const Day = ({ date }: DayProps) => {
-  return <section>{date}</section>;
+  return <section className={styles.day}>{date}</section>;
 };
 
 export default Day;
