+++ src/components/Month/Month.module.scss
@@ -1,0 +1,8 @@
+.month__heading {
+  margin-bottom: 32px;
+}
+
+.month__dayContainer {
+  display: flex;
+  flex-wrap: wrap;
+}
