+++ src/global.scss
@@ -18,21 +18,25 @@ a {
   color: #646cff;
   text-decoration: inherit;
 }
+
 a:hover {
   color: #535bf2;
 }
 
 body {
-  margin: 0;
   display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-}
+  flex-direction: column;
+  align-items: center;
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2.2rem;
+  text-align: center;
 
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
+  h1 {
+    margin-top: 0;
+    font-size: 3.2em;
+    line-height: 1.1;
+  }
 }
 
 button {
@@ -46,9 +50,11 @@ button {
   cursor: pointer;
   transition: border-color 0.25s;
 }
+
 button:hover {
   border-color: #646cff;
 }
+
 button:focus,
 button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
