+++ src/pages/DayDisplay/DayDisplay.tsx
@@ -1,9 +1,9 @@
 const DayDisplay = () => {
   return (
-    <>
+    <main>
       <h2>Day Display</h2>
       <p>Day display coming soon</p>
-    </>
+    </main>
   );
 };
