+++ src/pages/WeekDisplay/WeekDisplay.tsx
@@ -1,9 +1,9 @@
 const WeekDisplay = () => {
   return (
-    <>
+    <main>
       <h2>Week Display</h2>
       <p>Week display coming soon</p>
-    </>
+    </main>
   );
 };
