+++ src/pages/YearDisplay/YearDisplay.tsx
@@ -1,9 +1,9 @@
 const YearDisplay = () => {
   return (
-    <>
+    <main>
       <h2>Year Display</h2>
       <p>Year display coming soon</p>
-    </>
+    </main>
   );
 };
